# pledgecamp-oracle-develop/openapi.yaml
openapi: 3.0.0
info:
  title: Pledgecamp-Oracle
  version: v1.0
servers:
  - url: 'http://oracle.localdev.com:8081'
    description: Development
  - description: Production
    url: https://pledgecamp.com/api
...
paths:
  /projects/{project_id}:
    parameters:
      - schema:
          type: integer
        name: project_id
        in: path
        required: true
        description: Project ID from backend
    post:
      tags:
        - Project
      summary: ''
      operationId: post-projects-project_id
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
              examples: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
      description: Create a new project contract in a pre-initialized state. Call `SET_BACKERS` to initialize.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
                milestones:
                  type: array
                  items:
                    type: integer
                release_percents:
                  type: array
                  items:
                    type: integer
                creator:
                  type: integer
    get:
      tags:
        - Project
      summary: ''
      operationId: get-projects-project_id
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id:
                    type: integer
                  contract_address:
                    type: string
                  created_at:
                    type: string
                  modified_at:
                    type: string
                  completed_at:
                    type: string
                  next_activity_date:
                    type: string
                  activities_completed:
                    type: array
                    items:
                      type: string
                  project_activities_list:
                    type: object
                    properties:
                      project_activity_id:
                        type: integer
                      project_id:
                        type: integer
                      created_at:
                        type: string
                      modified_at:
                        type: string
                      transaction_hash:
                        type: object
                        properties:
                          string:
                            type: string
                          valid:
                            type: boolean
                      activity_status:
                        type: integer
                      activity_type:
                        type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
      description: Get information on the project state and the project's past activity
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
...